lidp_auditing/requirements.txt
absl-py~=1.4.0
dp-accounting ~= 0.4.3
numba ~= 0.57.1
numpy ~= 1.25.0
pandas ~= 1.5.3
scipy ~= 1.9.0
scikit-learn ~= 1.3.2
tensorboard ~= 2.14.0
tensorflow ~= 2.14.0
tensorflow_datasets ~= 4.9.3
